git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@16579 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/effect.c
@@ -2353,19 +2353,19 @@ MagickExport Image *KuwaharaImageChannel(const Image *image,
         {
           case 0:
           {
-            x_offset=x-width/2L-1;
-            y_offset=y-width/2L-1;
+            x_offset=x-(ssize_t) width/2L-1;
+            y_offset=y-(ssize_t) width/2L-1;
             break;
           }
           case 1:
           {
             x_offset=x;
-            y_offset=y-width/2L-1;
+            y_offset=y-(ssize_t) width/2L-1;
             break;
           }
           case 2:
           {
-            x_offset=x-width/2L-1;
+            x_offset=x-(ssize_t) width/2L-1;
             y_offset=y;
             break;
           }
